refactor(schools): move delete-confirmation into components
Removes the last legacy partials/ folder in the frontend. Refs #131.
