+++ resources/js/pages/schools/components/delete-confirmation.tsx
